move staff table into component and create table utilities file
